pyproject.toml
[project]
name = "apollo-miku-planner"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "matplotlib>=3.10.9",
    "numpy>=2.4.4",
    "osqp>=1.1.1",
    "scipy>=1.17.1",
]

[dependency-groups]
dev = [
    "python-docx>=1.2.0",
    "ruff>=0.15.12",
    "pytest>=8.0",
]

[tool.pytest.ini_options]
testpaths = ["工具/tests"]
